тестирование готово

+++ src/main/java/pro/sky/calculator/service/CalculatorServiceImpl.java
@@ -12,12 +12,12 @@ public class CalculatorServiceImpl implements CalculatorService {
 
     @Override
     public long plus(int num1, int num2) {
-        return num1 + num2;
+        return (long) num1 + num2;
     }
 
     @Override
     public long minus(int num1, int num2) {
-        return num1 - num2;
+        return (long) num1 - num2;
     }
     @Override
     public long multiply(int num1, int num2) {
